Move powerline setup into .zshrc
Somehow, it wasn't working in .zshenv-local. The powerline-daemon call
worked, but the command line definitely wasn't powerline. Might have to
do with the order in which .zshrc and .zshenv are parsed. If .zshenv is
first, it could be that it was clobbered by status line settings in
.zshrc.

Regardless, I've got a conditional for detecting Mac vs Linux, so I can
still use different paths to powerline, but check them into git so it's
not lost in the local-only .zshenv-local.