Install ALE, remove Syntastic
Syntastic has been commented out for a while in favor of pymode, but
it's time to upgrade to new hotness. ALE does the linting as you go, and
it's damn fast at it.

Also changed some pymode settings. I finally found the
pymode_rope_regenerate_on_write flag and shut it the hell off. I'll
probably need to look up how to manually kick off rope cache updates,
but not having to wait for them every time I save a file is a Good
Thing. Also going to try folding for a while, now that I've restyled it
a bit so it won't conflict with linter signs from ALE.

Hooray for Vim8 async!